move timedValue -> cachevalue package
